+++ config/jest.config.ts
@@ -6,7 +6,16 @@ export default async (): Promise<Config.InitialOptions> => ({
 
   rootDir: '../',
 
-  preset: 'ts-jest',
+  // if not using javascript files, use "ts-jest" instead
+  // preset: 'ts-jest',
+  preset: 'ts-jest/presets/js-with-ts-esm',
+
+  roots: [
+    '<rootDir>/src',
+    '<rootDir>/__tests__',
+  ],
+
+  coveragePathIgnorePatterns: ['.spec.util.ts',],
 
   // for DOM related:
   // testEnvironment: 'jsdom',
@@ -15,6 +24,10 @@ export default async (): Promise<Config.InitialOptions> => ({
   globals: {
     'ts-jest': {
       tsconfig: '<rootDir>/config/tsconfig.spec.json',
+
+      // if not using javscript files, remove the following:
+      useESM: true,
+      extensionsToTreatAsESM: ['.js', '.ts', '.tsx', '.jsx',],
     },
   },