+++ src/index.ts
@@ -2,4 +2,5 @@ export * from './flatten';
 export * from './write';
 export * from './nodes';
 export * from './parse';
+export * from './read';